+++ backend/internal/handlers/webhooks/webhooks.go
@@ -1,11 +1,11 @@
-package github
+package webhooks
 
 import (
 	models "github.com/CamPlume1/khoury-classroom/internal/models"
 	"github.com/gofiber/fiber/v2"
 )
 
-func (s *GitHubService) WebhookHandler(c *fiber.Ctx) error {
+func (s *WebHookService) WebhookHandler(c *fiber.Ctx) error {
 	var dispatch = map[string]func(c *fiber.Ctx) error{
 		"pull_request":                s.PR,
 		"pull_request_review_comment": s.PRComment,
@@ -21,12 +21,12 @@ func (s *GitHubService) WebhookHandler(c *fiber.Ctx) error {
 	return handler(c)
 }
 
-func (s *GitHubService) PR(c *fiber.Ctx) error {
+func (s *WebHookService) PR(c *fiber.Ctx) error {
 	println("PR webhook event")
 	return c.SendStatus(200)
 }
 
-func (s *GitHubService) PRComment(c *fiber.Ctx) error {
+func (s *WebHookService) PRComment(c *fiber.Ctx) error {
 	payload := models.PRComment{}
 	if err := c.BodyParser(&payload); err != nil {
 		return err
@@ -37,7 +37,7 @@ func (s *GitHubService) PRComment(c *fiber.Ctx) error {
 	return c.SendStatus(200)
 }
 
-func (s *GitHubService) PRThread(c *fiber.Ctx) error {
+func (s *WebHookService) PRThread(c *fiber.Ctx) error {
 	println("PR thread webhook event")
 	return c.SendStatus(200)
 }
